[13] Add support of annotation-based validations and split player name in tournament creation request

+++ tournament/tables/src/main/java/com/dss/tennis/tournament/tables/convertor/SequenceConverter.java
@@ -29,7 +29,7 @@ public class SequenceConverter<S, D extends AbstractSequentialDTO> extends Abstr
         return destination;
     }
 
-    private D convertSequenceElement(S sourceElement, long sequenceNumber) {
+    private D convertSequenceElement(S sourceElement, int sequenceNumber) {
         D destinationElement = modelMapper.map(sourceElement, this.destinationClass);
         destinationElement.setSequenceNumber(sequenceNumber);
